+++ pl_examples/basic_examples/lightning_module_template.py
@@ -1,7 +1,6 @@
 """
 Example template for defining a system
 """
-import logging as log
 import os
 from argparse import ArgumentParser
 from collections import OrderedDict
@@ -14,6 +13,7 @@ from torch import optim
 from torch.utils.data import DataLoader
 from torchvision.datasets import MNIST
 
+from pytorch_lightning import _logger as log
 from pytorch_lightning.core import LightningModule
 
 
@@ -225,7 +225,7 @@ class LightningTemplateModel(LightningModule):
         return self.__dataloader(train=False)
 
     @staticmethod
-    def add_model_specific_args(parent_parser, root_dir):  # pragma: no cover
+    def add_model_specific_args(parent_parser, root_dir):  # pragma: no-cover
         """
         Parameters you define here will be available to your model through self.hparams
         :param parent_parser: