+++ pytorch_lightning/__init__.py
@@ -1,6 +1,6 @@
 """Root package info."""
 
-__version__ = '0.7.1'
+__version__ = '0.7.2-dev'
 __author__ = 'William Falcon et al.'
 __author_email__ = 'waf2107@columbia.edu'
 __license__ = 'Apache-2.0'
@@ -19,15 +19,17 @@ except NameError:
     __LIGHTNING_SETUP__ = False
 
 if __LIGHTNING_SETUP__:
-    import sys
-    sys.stderr.write('Partial import of `torchlightning` during the build process.\n')
-    # We are not importing the rest of the scikit during the build
-    # process, as it may not be compiled yet
+    import sys  # pragma: no-cover
+    sys.stderr.write('Partial import of `torchlightning` during the build process.\n')  # pragma: no-cover
+    # We are not importing the rest of the lightning during the build process, as it may not be compiled yet
 else:
-    from .core import LightningModule
-    from .trainer import Trainer
-    from .callbacks import Callback
-    from .core import data_loader
+    from logging import getLogger
+    _logger = getLogger("lightning")
+
+    from pytorch_lightning.core import LightningModule
+    from pytorch_lightning.trainer import Trainer
+    from pytorch_lightning.callbacks import Callback
+    from pytorch_lightning.core import data_loader
 
     __all__ = [
         'Trainer',