+++ pytorch_lightning/loggers/base.py
@@ -4,6 +4,8 @@ from argparse import Namespace
 from functools import wraps
 from typing import Union, Optional, Dict, Iterable, Any, Callable, List
 
+import torch
+
 
 def rank_zero_only(fn: Callable):
     """Decorate a logger method to run it only on the process with rank 0.
@@ -30,7 +32,6 @@ class LightningLoggerBase(ABC):
     @abstractmethod
     def experiment(self) -> Any:
         """Return the experiment object associated with this logger"""
-        pass
 
     @abstractmethod
     def log_metrics(self, metrics: Dict[str, float], step: Optional[int] = None):
@@ -40,9 +41,9 @@ class LightningLoggerBase(ABC):
             metrics: Dictionary with metric names as keys and measured quantities as values
             step: Step number at which the metrics should be recorded
         """
-        pass
 
-    def _convert_params(self, params: Union[Dict[str, Any], Namespace]) -> Dict[str, Any]:
+    @staticmethod
+    def _convert_params(params: Union[Dict[str, Any], Namespace]) -> Dict[str, Any]:
         # in case converting from namespace
         if isinstance(params, Namespace):
             params = vars(params)
@@ -52,6 +53,62 @@ class LightningLoggerBase(ABC):
 
         return params
 
+    @staticmethod
+    def _flatten_dict(params: Dict[str, Any], delimiter: str = '/') -> Dict[str, Any]:
+        """Flatten hierarchical dict e.g. {'a': {'b': 'c'}} -> {'a/b': 'c'}.
+
+        Args:
+            params: Dictionary contains hparams
+            delimiter: Delimiter to express the hierarchy. Defaults to '/'.
+
+        Returns:
+            Flatten dict.
+
+        Examples:
+            >>> LightningLoggerBase._flatten_dict({'a': {'b': 'c'}})
+            {'a/b': 'c'}
+            >>> LightningLoggerBase._flatten_dict({'a': {'b': 123}})
+            {'a/b': 123}
+        """
+
+        def _dict_generator(input_dict, prefixes=None):
+            prefixes = prefixes[:] if prefixes else []
+            if isinstance(input_dict, dict):
+                for key, value in input_dict.items():
+                    if isinstance(value, (dict, Namespace)):
+                        value = vars(value) if isinstance(value, Namespace) else value
+                        for d in _dict_generator(value, prefixes + [key]):
+                            yield d
+                    else:
+                        yield prefixes + [key, value if value is not None else str(None)]
+            else:
+                yield prefixes + [input_dict if input_dict is None else str(input_dict)]
+
+        return {delimiter.join(keys): val for *keys, val in _dict_generator(params)}
+
+    @staticmethod
+    def _sanitize_params(params: Dict[str, Any]) -> Dict[str, Any]:
+        """Returns params with non-primitvies converted to strings for logging
+
+        >>> params = {"float": 0.3,
+        ...           "int": 1,
+        ...           "string": "abc",
+        ...           "bool": True,
+        ...           "list": [1, 2, 3],
+        ...           "namespace": Namespace(foo=3),
+        ...           "layer": torch.nn.BatchNorm1d}
+        >>> import pprint
+        >>> pprint.pprint(LightningLoggerBase._sanitize_params(params))  # doctest: +NORMALIZE_WHITESPACE
+        {'bool': True,
+         'float': 0.3,
+         'int': 1,
+         'layer': "<class 'torch.nn.modules.batchnorm.BatchNorm1d'>",
+         'list': '[1, 2, 3]',
+         'namespace': 'Namespace(foo=3)',
+         'string': 'abc'}
+        """
+        return {k: v if type(v) in [bool, int, float, str, torch.Tensor] else str(v) for k, v in params.items()}
+
     @abstractmethod
     def log_hyperparams(self, params: argparse.Namespace):
         """Record hyperparameters.
@@ -59,7 +116,6 @@ class LightningLoggerBase(ABC):
         Args:
             params: argparse.Namespace containing the hyperparameters
         """
-        pass
 
     def save(self) -> None:
         """Save log data."""
@@ -91,13 +147,11 @@ class LightningLoggerBase(ABC):
     @abstractmethod
     def name(self) -> str:
         """Return the experiment name."""
-        pass
 
     @property
     @abstractmethod
     def version(self) -> Union[int, str]:
         """Return the experiment version."""
-        pass
 
 
 class LoggerCollection(LightningLoggerBase):