+++ pytorch_lightning/loggers/tensorboard.py
@@ -8,7 +8,7 @@ import torch
 from pkg_resources import parse_version
 from torch.utils.tensorboard import SummaryWriter
 
-from .base import LightningLoggerBase, rank_zero_only
+from pytorch_lightning.loggers.base import LightningLoggerBase, rank_zero_only
 
 
 class TensorBoardLogger(LightningLoggerBase):
@@ -21,14 +21,12 @@ class TensorBoardLogger(LightningLoggerBase):
 
     .. _tf-logger:
 
-    Example
-    ------------------
+    Example:
+        .. code-block:: python
 
-    .. code-block:: python
-
-        logger = TensorBoardLogger("tb_logs", name="my_model")
-        trainer = Trainer(logger=logger)
-        trainer.train(model)
+            logger = TensorBoardLogger("tb_logs", name="my_model")
+            trainer = Trainer(logger=logger)
+            trainer.train(model)
 
     Args:
         save_dir (str): Save directory
@@ -101,6 +99,8 @@ class TensorBoardLogger(LightningLoggerBase):
     @rank_zero_only
     def log_hyperparams(self, params: Union[Dict[str, Any], Namespace]) -> None:
         params = self._convert_params(params)
+        params = self._flatten_dict(params)
+        sanitized_params = self._sanitize_params(params)
 
         if parse_version(torch.__version__) < parse_version("1.3.0"):
             warn(
@@ -110,13 +110,14 @@ class TensorBoardLogger(LightningLoggerBase):
             )
         else:
             from torch.utils.tensorboard.summary import hparams
-            exp, ssi, sei = hparams(params, {})
+            exp, ssi, sei = hparams(sanitized_params, {})
             writer = self.experiment._get_file_writer()
             writer.add_summary(exp)
             writer.add_summary(ssi)
             writer.add_summary(sei)
+
         # some alternative should be added
-        self.tags.update(params)
+        self.tags.update(sanitized_params)
 
     @rank_zero_only
     def log_metrics(self, metrics: Dict[str, float], step: Optional[int] = None) -> None: