+++ pytorch_lightning/loggers/test_tube.py
@@ -3,11 +3,11 @@ from typing import Optional, Dict, Any, Union
 
 try:
     from test_tube import Experiment
-except ImportError:
-    raise ImportError('You want to use `test_tube` logger which is not installed yet,'
+except ImportError:  # pragma: no-cover
+    raise ImportError('You want to use `test_tube` logger which is not installed yet,'  # pragma: no-cover
                       ' install it with `pip install test-tube`.')
 
-from .base import LightningLoggerBase, rank_zero_only
+from pytorch_lightning.loggers.base import LightningLoggerBase, rank_zero_only
 
 
 class TestTubeLogger(LightningLoggerBase):
@@ -96,6 +96,7 @@ class TestTubeLogger(LightningLoggerBase):
         # TODO: HACK figure out where this is being set to true
         self.experiment.debug = self.debug
         params = self._convert_params(params)
+        params = self._flatten_dict(params)
         self.experiment.argparse(Namespace(**params))
 
     @rank_zero_only