+++ tests/models/base.py
@@ -183,7 +183,7 @@ class TestModelBase(LightningModule):
         return loader
 
     @staticmethod
-    def add_model_specific_args(parent_parser, root_dir):  # pragma: no cover
+    def add_model_specific_args(parent_parser, root_dir):  # pragma: no-cover
         """
         Parameters you define here will be available to your model through self.hparams
         :param parent_parser: